Impl: LSTM with functional API

+++ readers/AbstractProcessLogReader.py
@@ -174,7 +174,7 @@ class AbstractProcessLogReader():
         return tf.data.Dataset.from_generator(
             self._generate_examples,
             args=['train'],
-            output_types=(tf.int64, tf.int64),
+            output_types=(tf.float32, tf.float32),
             output_shapes=((None, ), (
                 None,
                 None,
